Updated the solution file and project files to Visual Studio 2010. Cleaned up some code. Changed the connection succeeded message. Removed the client resource download messages and fixed the file transfer class. Added patch to disable random ambient animations. Renamed CPadState class to CControlState. Fixed CIVPad custom constructor not setting an initial value for the unknown member of CPadData (Fixes player ped control states not working properly). Added a load more task names and types to the CIVTask class. Added CIVTask::GetParent and CIVTask::GetSubTask. Fixed some includes. Fixed death detection not working (See CNetworkPlayer::IsDead hack). Enabled new vehicle creation and destruction code. Moved the DirectX SDK to the Vendor directory. Added local player debug view code. Renamed getPlayerPadState, getPlayerPreviousPadState and getEmptyPadState natives to getPlayerControlState, getPlayerPreviousControlState and getEmptyControlState (Both clientside and serverside, old native names also kept for backwards compatibility). Fixed connection packet being processed twice (Missing return from CNetClient::ProcessPacket and incorrect INTERNAL_PACKET_END macro). Removed gcc GetTickCount from SharedUtility (SharedUtility::GetTime is used now). Incremented the network and network module version. Moved DEBUG_SUFFIX macro to Common.h.

+++ trunk/Client/Core/CEGUI/cegui/src/CEGUIDefaultResourceProvider.cpp
@@ -65,7 +65,7 @@ namespace CEGUI
         std::streampos size = dataFile.tellg();
         dataFile.seekg (0, std::ios::beg);
 
-        unsigned char* buffer = new unsigned char [size];
+        unsigned char* buffer = new unsigned char [static_cast<unsigned int>(size)];
 
         try {
             dataFile.read(reinterpret_cast<char*>(buffer), size);
@@ -79,7 +79,7 @@ namespace CEGUI
         dataFile.close();
 
         output.setData(buffer);
-        output.setSize(size);
+        output.setSize(static_cast<size_t>(size));
     }
     
     void DefaultResourceProvider::unloadRawDataContainer(RawDataContainer& data)
